refactored parallel code to be split among utility, algorithm, and main files
